requirements.txt
#django-autoslug
dcg-autoslug

django-dept-directory
django-markuphelpers
